backend/requirements.txt
flask==2.0.1
flask-cors==3.0.10
pymongo==4.0.1
redis==4.0.2
kafka-python==2.0.2
python-dotenv==0.19.1
flask-jwt-extended==4.3.1
gunicorn==20.1.0
bcrypt==3.2.0
